refactor: rewrite Guild events
