+++ .gitignore
@@ -4,10 +4,11 @@ Thumbs.db
 
 # Configuration
 /config.json
-/config.bak.json
+/config.json.*
 
 # Node
 node_modules
-dist
 
+# TypeScript build
+dist
 tsconfig.tsbuildinfo
