+++ config.sample.json
@@ -14,6 +14,9 @@
 		]
 	},
 	"devGuildsIds": [],
+	"devLogs": {
+		"joins": ""
+	},
 	"embed": {
 		"color": "#00FF00",
 		"footer": {
